# [WAN] Use different sharding strategy for self and cross attention.